+++ src/ray.rs
@@ -4,17 +4,17 @@ use crate::vec3::*;
 pub struct Ray {
     pub origin: Point3,
     pub direction: Vec3,
-    pub time: f64
+    pub time: f32
 }
 
 impl Ray {
-    pub fn new(origin: Point3, direction: Vec3, time: f64) -> Ray {
+    pub fn new(origin: Point3, direction: Vec3, time: f32) -> Ray {
         Ray {
             origin, direction, time
         }
     }
 
-    pub fn at(self, t: f64) -> Point3 {
+    pub fn at(self, t: f32) -> Point3 {
         self.origin + t * self.direction
     }
 }
